+++ src/Options/Number.php
@@ -12,4 +12,17 @@ class Number extends OptionAbstract
 
         parent::__construct($args, $meta_box);
     }
+
+    public function get_input_attributes_string($attributes = [])
+    {
+        if ($this->get_arg('min') !== null) {
+            $attributes['min'] = $this->get_arg('min');
+        }
+
+        if ($this->get_arg('max') !== null) {
+            $attributes['max'] = $this->get_arg('max');
+        }
+
+        return parent::get_input_attributes_string($attributes);
+    }
 }